feat: Add recipient to arrival notification email
This commit adds an additional recipient to the arrival notification email in the `SendEmailForArrivalCandidates` job. The email will now be sent to both 'sasocvetanoski@gmail.com' and 'sase_cvetanoski@hotmail.com'. This change ensures that both recipients receive the arrival notification.

In addition, the `arrivalCandidateWithStatus.blade.php` view has been updated to include a link to add the arrival event to the Google Calendar. A new link has also been added to download an Outlook Calendar event file (.ics) for the arrival event.

These changes improve the functionality and usability of the arrival notification feature.