feat: Add passwordShow field to users table
This commit adds a new field, passwordShow, to the users table. The passwordShow field is nullable and will be used to store the plain text password temporarily during the password change process. This change enhances the functionality of the ChangePasswordForUserCommand by allowing it to store the new password in plain text before hashing it with bcrypt and updating the user's password field.