Update ArrivalCandidateController to update candidate status based on arrival status
This commit updates the ArrivalCandidateController to include logic for updating the candidate status based on the arrival status. The candidate status is determined based on the value of the status_arrival_id field in the arrival candidate. If the status_arrival_id is 1, the candidate status is set to 5 (Pristignal). If the status_arrival_id is 3, the candidate status is set to 6 (Procedura za ERPR). If the status_arrival_id is 4, the candidate status is set to 22 (Procedura za pismo). If the status_arrival_id is 5, the candidate status is set to 7 (Snimka za ERPR). If the status_arrival_id is 6, the candidate status is set to 8 (Poluchava ERPR). If the status_arrival_id is 9, the candidate status is set to 9 (Naznachen za rabota). This change improves the functionality of the application by automatically updating the candidate status based on the arrival status.