feat: Add conditionsMetDeclaration, jobDescription, and employmentContract fields to MigrationDocumentPreparation
This commit adds the fields conditionsMetDeclaration, jobDescription, and employmentContract to the MigrationDocumentPreparation model and its corresponding migrations. These fields are used to store additional information related to the preparation of migration documents. The purpose of this change is to enhance the functionality of the MigrationDocumentPreparation model by including these new fields.