feat: Update LoginController index method to filter users by role_id
This commit modifies the index method in the LoginController.php file to include a filter for users based on the role_id parameter. If the role_id is provided in the request, the method will retrieve users with the specified role_id. If no role_id is provided, the method will retrieve all users except for the one with id 22. This change improves the flexibility of the index method by allowing users to filter the results based on their role.